use tidb_max_tiflash_threads as stream count

+++ planner/core/plan.go
@@ -429,7 +429,7 @@ type basePhysicalPlan struct {
 	planCostInit bool
 	planCost     float64
 
-	TiFlashFineGrainedShuffleStreamCount uint32
+	TiFlashFineGrainedShuffleStreamCount int64
 }
 
 // Cost implements PhysicalPlan interface.
